fixing issue with styles.css, reverting package.json
styles.css was being referenced in 2 modules as Styles.css. Changing those refs to be correct (i.e. to styles.css) somehow breaks the application, while leaving them to be incorrectly capitalized works just fine. Renaming the file to be capitalized works.  Thinking the Capitalized filename, while wrong, is somehow compiled into the "solution" such that changing the refs to be correct causes them to refer to something that isn't there. But I can't seem to rebuild the thing in such a way as to get the refs in sync. thus, just renaming the source css file to match the refs.